perf: 修改 GatewaySerializer 目录
