+++ apps/assets/serializers/__init__.py
@@ -4,11 +4,11 @@
 from .asset import *
 from .label import *
 from .node import *
+from .gateway import *
 from .domain import *
 from .gathered_user import *
 from .favorite_asset import *
 from .account import *
-from assets.serializers.account.backup import *
 from .platform import *
 from .cagegory import *
 from .automations import *
