+++ apps/assets/serializers/account/__init__.py
@@ -1,2 +1,3 @@
 from .account import *
 from .template import *
+from .backup import *
